Address Copilot review on #134
Report a canvas_snap failure from snaptile as ERR_OUT_OF_SPACE, matching
snapsh, instead of returning quietly with the slot unfilled. Both shipped
canvas_snap implementations return true unconditionally, so this is a
contract fix rather than a reachable bug.

Drop a pointless (uint16_t) narrowing of Token.length, which is size_t.
The overflow Copilot described cannot happen -- a blob word is never a
list element -- but the cast said otherwise. Noted the invariant in
copilot-instructions.md.

+++ core/token_source.c
@@ -176,7 +176,7 @@ static uint8_t word_view(Node element, const char **str, size_t *len)
 static Token token_from_class(uint8_t cls, const char *str, size_t len,
                               bool prev_was_delimiter)
 {
-    Token t = {.type = TOKEN_WORD, .start = str, .length = (uint16_t)len};
+    Token t = {.type = TOKEN_WORD, .start = str, .length = len};
 
     if (cls != ATOM_CLASS_CONTEXT)
     {
